+++ webui/src/components/MonacoEditor/index.js
@@ -2,7 +2,6 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { noop, throttle } from 'lodash';
 import monaco from '../../misc/initMonacoEditor';
-import { subscribeOnTargetEvent } from '../../misc/eventHandler';
 import { getModelByFile, setModelByFile } from '../../misc/monacoModelStorage';
 import { getYAMLError } from '../../misc/yamlValidation';
 
@@ -55,7 +54,6 @@ export default class MonacoEditor extends React.Component {
 
   componentDidMount() {
     this.initMonaco();
-    this.unsubcribeResize = subscribeOnTargetEvent(window, 'resize', this.throttledAdjustEditorSize)
   }
 
   adjustEditorSize() {
@@ -63,17 +61,13 @@ export default class MonacoEditor extends React.Component {
       const { width, height } = this.state
       const { clientWidth, clientHeight } = this.containerElement
       if (width !== clientWidth && height !== clientHeight) {
-        this.setState(() => ({
+        this.setState({
           height: clientHeight,
           width: clientWidth
-        }), () => {
-          this.editor.layout()
         })
       }
     }
   }
-
-  throttledAdjustEditorSize = throttle(this.adjustEditorSize, 1000, { leading: false })
 
   setValidationError = () => {
     const {
@@ -127,7 +121,7 @@ export default class MonacoEditor extends React.Component {
     if (prevProps.theme !== theme) {
       monaco.editor.setTheme(theme);
     }
-    this.throttledAdjustEditorSize()
+
     if (prevProps.options !== options) {
       editor.updateOptions(options);
     }
@@ -139,7 +133,6 @@ export default class MonacoEditor extends React.Component {
 
   componentWillUnmount() {
     this.destroyMonaco();
-    this.unsubcribeResize();
   }
 
   assignRef = component => {