+++ webui/src/components/SchemaEditor/index.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { noop, throttle } from 'lodash';
 import monaco from '../../misc/initMonacoEditor';
-import { subscribeOnTargetEvent } from '../../misc/eventHandler';
 import { getYAMLError } from '../../misc/yamlValidation';
 
 const containerStyles = css`
@@ -55,27 +54,7 @@ export default class SchemaEditor extends React.Component {
 
   componentDidMount() {
     this.initMonaco();
-    this.unsubcribeResize = subscribeOnTargetEvent(window, 'resize', this.throttledAdjustEditorSize)
   }
-
-  adjustEditorSize = () => {
-    if (this.containerElement && this.containerElement.current && this.editor) {
-      const { width, height } = this.state;
-      const { clientWidth, clientHeight } = this.containerElement.current;
-
-      if (width !== clientWidth || height !== clientHeight) {
-        this.setState(
-          {
-            height: clientHeight,
-            width: clientWidth
-          },
-          () => this.editor.layout()
-        );
-      }
-    }
-  }
-
-  throttledAdjustEditorSize = throttle(this.adjustEditorSize, 1000, { leading: false })
 
   componentDidUpdate(prevProps) {
     const {
@@ -109,7 +88,6 @@ export default class SchemaEditor extends React.Component {
     if (prevProps.theme !== theme) {
       monaco.editor.setTheme(theme);
     }
-    this.throttledAdjustEditorSize()
     if (prevProps.options !== options) {
       editor.updateOptions(options);
     }
@@ -123,7 +101,6 @@ export default class SchemaEditor extends React.Component {
 
   componentWillUnmount() {
     this.destroyMonaco();
-    this.unsubcribeResize();
   }
 
   destroyMonaco() {
@@ -185,8 +162,6 @@ export default class SchemaEditor extends React.Component {
 
   editorDidMount(editor) {
     this.props.editorDidMount(editor, monaco);
-
-    this.adjustEditorSize();
 
     if (this.props.cursor !== DEF_CURSOR) {
       editor.setSelection(this.props.cursor)